# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: STACK Service Desk AI Solution API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets (role-filtered)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: use_case
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: sla_status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketListResponse"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a new ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /v1/tickets/{ticketId}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket by ID
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    delete:
      operationId: deleteTicket
      tags: [tickets]
      summary: Delete ticket (admin only)
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /v1/tickets/{ticketId}/resolve:
    post:
      operationId: resolveTicket
      tags: [tickets]
      summary: Trigger AI resolution
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIResolution"
  /v1/tickets/{ticketId}/escalate:
    post:
      operationId: escalateTicket
      tags: [tickets]
      summary: Escalate ticket to agent
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Escalated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /v1/tickets/{ticketId}/notes:
    get:
      operationId: listTicketNotes
      tags: [tickets]
      summary: Get ticket notes
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketNote"
    post:
      operationId: addTicketNote
      tags: [tickets]
      summary: Add note to ticket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketNote"
  /v1/tickets/{ticketId}/timeline:
    get:
      operationId: getTicketTimeline
      tags: [tickets]
      summary: Get ticket audit timeline
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEvent"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Ticket:
      type: object
      required: [ticket_id, title, use_case, status, priority, source, sla_status, created_at]
      properties:
        ticket_id:
          type: string
        freshservice_ticket_id:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        use_case:
          type: string
        status:
          type: string
        priority:
          type: string
        sla_deadline:
          type: ["string", "null"]
        sla_status:
          type: string
        sla_breach_predicted:
          type: boolean
        source:
          type: string
        user_id:
          type: string
        assigned_agent_id:
          type: ["string", "null"]
        resolution_type:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
        closed_at:
          type: ["string", "null"]
        confidence_score:
          type: ["number", "null"]
        user_name:
          type: ["string", "null"]
        agent_name:
          type: ["string", "null"]
    TicketDetail:
      type: object
      required: [ticket_id, title, use_case, status, priority, source, sla_status, created_at]
      properties:
        ticket_id:
          type: string
        freshservice_ticket_id:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        use_case:
          type: string
        status:
          type: string
        priority:
          type: string
        sla_deadline:
          type: ["string", "null"]
        sla_status:
          type: string
        sla_breach_predicted:
          type: boolean
        source:
          type: string
        user_id:
          type: string
        assigned_agent_id:
          type: ["string", "null"]
        resolution_type:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
        closed_at:
          type: ["string", "null"]
        confidence_score:
          type: ["number", "null"]
        user_name:
          type: ["string", "null"]
        agent_name:
          type: ["string", "null"]
        ai_resolution:
          $ref: "#/components/schemas/AIResolution"
    TicketListResponse:
      type: object
      required: [tickets, total, page, limit]
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TicketInput:
      type: object
      required: [title, use_case, source]
      properties:
        title:
          type: string
        description:
          type: string
        use_case:
          type: string
        priority:
          type: string
        source:
          type: string
        device_name:
          type: string
        metadata:
          type: object
          additionalProperties: true
    TicketUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assigned_agent_id:
          type: string
    TicketNote:
      type: object
      required: [note_id, ticket_id, note_type, content, created_at]
      properties:
        note_id:
          type: string
        ticket_id:
          type: string
        note_type:
          type: string
        content:
          type: string
        created_by:
          type: ["string", "null"]
        is_internal:
          type: boolean
        created_at:
          type: string
        author_name:
          type: ["string", "null"]
    NoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        note_type:
          type: string
        is_internal:
          type: boolean
    AIResolution:
      type: object
      required: [resolution_id, ticket_id, decision, confidence_score, created_at]
      properties:
        resolution_id:
          type: string
        ticket_id:
          type: string
        intent_detected:
          type: ["string", "null"]
        root_cause:
          type: ["string", "null"]
        confidence_score:
          type: number
        intent_clarity_score:
          type: ["number", "null"]
        sop_match_score:
          type: ["number", "null"]
        historical_success_score:
          type: ["number", "null"]
        input_completeness_score:
          type: ["number", "null"]
        decision:
          type: string
        resolution_steps:
          type: ["object", "null"]
          additionalProperties: true
        execution_status:
          type: ["string", "null"]
        execution_output:
          type: ["string", "null"]
        time_taken_seconds:
          type: ["integer", "null"]
        sop_title:
          type: ["string", "null"]
        created_at:
          type: string
    AuditEvent:
      type: object
      required: [log_id, event_type, actor_type, created_at]
      properties:
        log_id:
          type: string
        ticket_id:
          type: ["string", "null"]
        event_type:
          type: string
        actor_id:
          type: ["string", "null"]
        actor_type:
          type: string
        actor_name:
          type: ["string", "null"]
        details:
          type: ["object", "null"]
          additionalProperties: true
        ip_address:
          type: ["string", "null"]
        platform:
          type: ["string", "null"]
        created_at:
          type: string
...